Store last open file path for future sessions
