--- java/bundles/org.eclipse.set.core.rcp.dialogService/.settings/org.eclipse.core.resources.prefs
@@ -1,2 +1,0 @@
-eclipse.preferences.version=1
-encoding/<project>=UTF-8
