--- java/bundles/org.eclipse.set.core.rcp.dialogService/.settings/org.eclipse.jdt.ui.prefs
@@ -1,107 +1,0 @@
-eclipse.preferences.version=1
-editor_save_participant_org.eclipse.jdt.ui.postsavelistener.cleanup=true
-formatter_profile=_PlanPro
-formatter_settings_version=20
-org.eclipse.jdt.ui.text.custom_code_templates=
-sp_cleanup.add_all=false
-sp_cleanup.add_default_serial_version_id=true
-sp_cleanup.add_generated_serial_version_id=false
-sp_cleanup.add_missing_annotations=true
-sp_cleanup.add_missing_deprecated_annotations=true
-sp_cleanup.add_missing_methods=false
-sp_cleanup.add_missing_nls_tags=false
-sp_cleanup.add_missing_override_annotations=true
-sp_cleanup.add_missing_override_annotations_interface_methods=true
-sp_cleanup.add_serial_version_id=false
-sp_cleanup.always_use_blocks=true
-sp_cleanup.always_use_parentheses_in_expressions=false
-sp_cleanup.always_use_this_for_non_static_field_access=false
-sp_cleanup.always_use_this_for_non_static_method_access=false
-sp_cleanup.arrays_fill=false
-sp_cleanup.bitwise_conditional_expression=false
-sp_cleanup.boolean_literal=false
-sp_cleanup.break_loop=false
-sp_cleanup.collection_cloning=false
-sp_cleanup.comparison_statement=false
-sp_cleanup.convert_functional_interfaces=false
-sp_cleanup.convert_to_enhanced_for_loop=true
-sp_cleanup.convert_to_enhanced_for_loop_if_loop_var_used=false
-sp_cleanup.convert_to_switch_expressions=false
-sp_cleanup.correct_indentation=false
-sp_cleanup.double_negation=false
-sp_cleanup.else_if=false
-sp_cleanup.embedded_if=false
-sp_cleanup.evaluate_nullable=false
-sp_cleanup.format_source_code=true
-sp_cleanup.format_source_code_changes_only=false
-sp_cleanup.hash=false
-sp_cleanup.if_condition=false
-sp_cleanup.insert_inferred_type_arguments=false
-sp_cleanup.instanceof=false
-sp_cleanup.join=false
-sp_cleanup.lazy_logical_operator=false
-sp_cleanup.make_local_variable_final=true
-sp_cleanup.make_parameters_final=true
-sp_cleanup.make_private_fields_final=true
-sp_cleanup.make_type_abstract_if_missing_method=false
-sp_cleanup.make_variable_declarations_final=true
-sp_cleanup.map_cloning=false
-sp_cleanup.merge_conditional_blocks=false
-sp_cleanup.never_use_blocks=false
-sp_cleanup.never_use_parentheses_in_expressions=true
-sp_cleanup.no_string_creation=false
-sp_cleanup.no_super=false
-sp_cleanup.number_suffix=false
-sp_cleanup.objects_equals=false
-sp_cleanup.on_save_use_additional_actions=true
-sp_cleanup.organize_imports=true
-sp_cleanup.overridden_assignment=false
-sp_cleanup.precompile_regex=false
-sp_cleanup.primitive_serialization=false
-sp_cleanup.pull_up_assignment=false
-sp_cleanup.push_down_negation=false
-sp_cleanup.qualify_static_field_accesses_with_declaring_class=false
-sp_cleanup.qualify_static_member_accesses_through_instances_with_declaring_class=true
-sp_cleanup.qualify_static_member_accesses_through_subtypes_with_declaring_class=true
-sp_cleanup.qualify_static_member_accesses_with_declaring_class=true
-sp_cleanup.qualify_static_method_accesses_with_declaring_class=false
-sp_cleanup.redundant_falling_through_block_end=false
-sp_cleanup.remove_private_constructors=true
-sp_cleanup.remove_redundant_modifiers=false
-sp_cleanup.remove_redundant_semicolons=false
-sp_cleanup.remove_redundant_type_arguments=true
-sp_cleanup.remove_trailing_whitespaces=false
-sp_cleanup.remove_trailing_whitespaces_all=true
-sp_cleanup.remove_trailing_whitespaces_ignore_empty=false
-sp_cleanup.remove_unnecessary_array_creation=false
-sp_cleanup.remove_unnecessary_casts=true
-sp_cleanup.remove_unnecessary_nls_tags=true
-sp_cleanup.remove_unused_imports=false
-sp_cleanup.remove_unused_local_variables=false
-sp_cleanup.remove_unused_private_fields=true
-sp_cleanup.remove_unused_private_members=false
-sp_cleanup.remove_unused_private_methods=true
-sp_cleanup.remove_unused_private_types=true
-sp_cleanup.simplify_lambda_expression_and_method_ref=false
-sp_cleanup.sort_members=false
-sp_cleanup.sort_members_all=true
-sp_cleanup.strictly_equal_or_different=false
-sp_cleanup.stringbuilder=false
-sp_cleanup.switch=false
-sp_cleanup.ternary_operator=false
-sp_cleanup.try_with_resource=false
-sp_cleanup.use_anonymous_class_creation=false
-sp_cleanup.use_autoboxing=false
-sp_cleanup.use_blocks=true
-sp_cleanup.use_blocks_only_for_return_and_throw=false
-sp_cleanup.use_directly_map_method=false
-sp_cleanup.use_lambda=true
-sp_cleanup.use_parentheses_in_expressions=true
-sp_cleanup.use_this_for_non_static_field_access=false
-sp_cleanup.use_this_for_non_static_field_access_only_if_necessary=true
-sp_cleanup.use_this_for_non_static_method_access=false
-sp_cleanup.use_this_for_non_static_method_access_only_if_necessary=true
-sp_cleanup.use_unboxing=false
-sp_cleanup.use_var=false
-sp_cleanup.useless_continue=false
-sp_cleanup.useless_return=false
