--- java/bundles/org.eclipse.set.core.rcp.dialogService/.settings/org.eclipse.pde.ds.annotations.prefs
@@ -1,8 +1,0 @@
-classpath=true
-dsVersion=V1_3
-eclipse.preferences.version=1
-enabled=true
-generateBundleActivationPolicyLazy=true
-path=OSGI-INF
-validationErrorLevel=error
-validationErrorLevel.missingImplicitUnbindMethod=error
