--- java/bundles/org.eclipse.set.core.rcp.dialogService/.settings/org.eclipse.wst.sse.core.prefs
@@ -1,4 +1,0 @@
-eclipse.preferences.version=1
-task-tags/taskPriorities=1,2,1,0
-task-tags/taskTags=TODO,FIXME,XXX,RAP_IMPROVE
-task-tags/use-project-settings=true
