--- java/bundles/org.eclipse.set.core.rcp.dialogService/.settings/org.eclipse.xtend.core.Xtend.prefs
@@ -1,80 +1,0 @@
-BuilderConfiguration.is_project_specific=true
-ValidatorConfiguration.is_project_specific=true
-autobuilding=true
-blank.line.around.expressions=0
-blank.lines.after.imports=1
-blank.lines.after.last.member=0
-blank.lines.after.packagedecl=1
-blank.lines.before.first.member=0
-blank.lines.between.classes=1
-blank.lines.between.enum.literals=0
-blank.lines.between.fields=0
-blank.lines.between.fields.and.methods=1
-blank.lines.between.imports=0
-blank.lines.between.methods=1
-braces.in.new.line=false
-eclipse.preferences.version=1
-generateGeneratedAnnotation=false
-generateSuppressWarnings=true
-generatedAnnotationComment=
-includeDateInGenerated=false
-indentation=\t
-indentation.length=4
-keep.one.line.methods=true
-line.separator=\r\n
-line.width.max=80
-newline.after.class.annotations=false
-newline.after.constructor.annotations=false
-newline.after.field.annotations=false
-newline.after.method.annotations=false
-newline.after.parameter.annotations=false
-org.eclipse.xtend.core.Xtend.useProjectSettings=true
-org.eclipse.xtend.core.validation.IssueCodes.api_type_inference=warning
-org.eclipse.xtend.core.validation.IssueCodes.dispatch_plain_function_name_clash=warning
-org.eclipse.xtend.core.validation.IssueCodes.implicit_return=ignore
-org.eclipse.xtend.core.validation.IssueCodes.invalid_operator_signature=error
-org.eclipse.xtend.core.validation.IssueCodes.java_doc_linking=org.eclipse.jdt.core.compiler.problem.invalidJavadoc\:ignore
-org.eclipse.xtend.core.validation.IssueCodes.orphan_element=warning
-org.eclipse.xtend.core.validation.IssueCodes.single_case_function=warning
-org.eclipse.xtend.core.validation.IssueCodes.unused_private_member=org.eclipse.jdt.core.compiler.problem.unusedPrivateMember\:warning
-org.eclipse.xtend.formatter.version=1
-org.eclipse.xtend.ide.default.eclipse_profile=_PlanPro
-org.eclipse.xtext.builder.copyJavaProblems=error
-org.eclipse.xtext.xbase.validation.IssueCodes.constant_condition=warning
-org.eclipse.xtext.xbase.validation.IssueCodes.deprecated_member_reference=org.eclipse.jdt.core.compiler.problem.deprecation\:warning
-org.eclipse.xtext.xbase.validation.IssueCodes.discouraged_reference=org.eclipse.jdt.core.compiler.problem.discouragedReference\:warning
-org.eclipse.xtext.xbase.validation.IssueCodes.equals_with_null=warning
-org.eclipse.xtext.xbase.validation.IssueCodes.forbidden_reference=org.eclipse.jdt.core.compiler.problem.forbiddenReference\:warning
-org.eclipse.xtext.xbase.validation.IssueCodes.import_unsued=org.eclipse.jdt.core.compiler.problem.unusedImport\:warning
-org.eclipse.xtext.xbase.validation.IssueCodes.import_wildcard_deprecated=warning
-org.eclipse.xtext.xbase.validation.IssueCodes.incomplete_cases_on_enum=warning
-org.eclipse.xtext.xbase.validation.IssueCodes.java_style_type_cast=error
-org.eclipse.xtext.xbase.validation.IssueCodes.null_safe_feature_call_on_primitive_valued_feature=warning
-org.eclipse.xtext.xbase.validation.IssueCodes.obsolete_cast=org.eclipse.jdt.core.compiler.problem.unnecessaryTypeCheck\:warning
-org.eclipse.xtext.xbase.validation.IssueCodes.obsolete_instanceof=org.eclipse.jdt.core.compiler.problem.unnecessaryTypeCheck\:warning
-org.eclipse.xtext.xbase.validation.IssueCodes.operation_without_parentheses=ignore
-org.eclipse.xtext.xbase.validation.IssueCodes.suspiciously_overloaded_feature=warning
-org.eclipse.xtext.xbase.validation.IssueCodes.unhandled_exception=warning
-org.eclipse.xtext.xbase.validation.IssueCodes.unreachable_case=error
-org.eclipse.xtext.xbase.validation.IssueCodes.unreachable_instance_of=error
-org.eclipse.xtext.xbase.validation.IssueCodes.unused_local_variable=org.eclipse.jdt.core.compiler.problem.unusedLocal\:warning
-org.eclipse.xtext.xbase.validation.IssueCodes.variable_name_discouraged=warning
-outlet.DEFAULT_OUTPUT.cleanDirectory=false
-outlet.DEFAULT_OUTPUT.cleanupDerived=true
-outlet.DEFAULT_OUTPUT.createDirectory=true
-outlet.DEFAULT_OUTPUT.derived=true
-outlet.DEFAULT_OUTPUT.directory=xtend-gen
-outlet.DEFAULT_OUTPUT.hideLocalSyntheticVariables=true
-outlet.DEFAULT_OUTPUT.installDslAsPrimarySource=false
-outlet.DEFAULT_OUTPUT.keepLocalHistory=false
-outlet.DEFAULT_OUTPUT.override=true
-outlet.DEFAULT_OUTPUT.sourceFolder.src.directory=
-outlet.DEFAULT_OUTPUT.sourceFolder.src.ignore=
-outlet.DEFAULT_OUTPUT.sourceFolder.xtend-gen.directory=
-outlet.DEFAULT_OUTPUT.sourceFolder.xtend-gen.ignore=
-outlet.DEFAULT_OUTPUT.userOutputPerSourceFolder=
-preserve.blank.lines=1
-preserve.new.lines=true
-targetJavaVersion=JAVA17
-useJavaCompilerCompliance=true
-whitespace.between.keyword.and.parenthesis=true
