--- java/bundles/org.eclipse.set.core.rcp.dialogService/META-INF/MANIFEST.MF
@@ -1,32 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-Name: Toolbox RCP DialogService
-Bundle-SymbolicName: org.eclipse.set.core.rcp.dialogservice
-Bundle-Version: 2.3.0.qualifier
-Bundle-Vendor: Eclipse Signalling Engineering Toolbox
-Automatic-Module-Name: org.eclipse.set.core.rcp.dialogservice
-Bundle-RequiredExecutionEnvironment: JavaSE-21
-Bundle-ActivationPolicy: lazy
-Import-Package: com.google.common.collect;version="15.0.0",
- org.eclipse.set.basis,
- org.eclipse.set.basis.extensions,
- org.eclipse.set.basis.files,
- org.eclipse.set.core,
- org.eclipse.set.core.dialogservice,
- org.eclipse.set.core.services.dialog,
- org.eclipse.set.custom.extensions,
- org.eclipse.set.utils,
- org.eclipse.set.utils.widgets,
- jakarta.inject,
- org.eclipse.e4.core.contexts,
- org.eclipse.e4.core.di,
- org.eclipse.e4.core.services.nls,
- org.eclipse.e4.ui.di,
- org.eclipse.e4.ui.model,
- org.eclipse.e4.ui.model.application,
- org.eclipse.e4.ui.model.application.ui,
- org.eclipse.swt.widgets,
- org.osgi.service.component.annotations,
- org.slf4j
-Require-Bundle: org.eclipse.set.rcp.compatibility;bundle-version="0.1.0"
-Service-Component: OSGI-INF/org.eclipse.set.core.rcp.dialogservice.DialogServiceContextFunction.xml
