--- java/bundles/org.eclipse.set.core.rcp.dialogService/build.properties
@@ -1,5 +1,0 @@
-source.. = src/
-output.. = bin/
-bin.includes = META-INF/,\
-               .,\
-               OSGI-INF/
