--- java/bundles/org.eclipse.set.core.rcp.dialogService/pom.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-
-	<groupId>org.eclipse.set</groupId>
-	<artifactId>org.eclipse.set.core.rcp.dialogservice</artifactId>
-	<version>2.3.0-SNAPSHOT</version>
-	<packaging>eclipse-plugin</packaging>
-
-	<parent>
-		<relativePath>../../..</relativePath>
-		<groupId>org.eclipse.set</groupId>
-		<artifactId>org.eclipse.set.releng.parent</artifactId>
-		<version>2.3.0-SNAPSHOT</version>
-	</parent>
-</project>
