+++ java/bundles/org.eclipse.set.core/META-INF/MANIFEST.MF
@@ -79,6 +79,7 @@ Import-Package: com.fasterxml.jackson.annotation,
  org.eclipse.set.core.services.validation,
  org.eclipse.set.core.services.version,
  org.eclipse.set.core.services.viewmodel,
+ org.eclipse.set.custom.extensions,
  org.eclipse.set.model.validationreport,
  org.eclipse.set.model.zipmanifest,
  org.eclipse.set.ppmodel.extensions,
@@ -112,6 +113,7 @@ Export-Package: org.eclipse.set.core,
  org.eclipse.set.core.modelservice
 Service-Component: OSGI-INF/org.eclipse.set.core.Messages.xml,
  OSGI-INF/org.eclipse.set.core.configurationservice.UserConfigurationServiceImpl.xml,
+ OSGI-INF/org.eclipse.set.core.dialogservice.DialogServiceContextFunction.xml,
  OSGI-INF/org.eclipse.set.core.enumtranslation.EnumTranslationServiceImpl.xml,
  OSGI-INF/org.eclipse.set.core.enumtranslation.Enumerators.xml,
  OSGI-INF/org.eclipse.set.core.fileservice.PlainFileFormatService.xml,
