+++ java/bundles/org.eclipse.set.swtbot/src/org/eclipse/set/swtbot/utils/MockDialogService.java
@@ -131,11 +131,6 @@ public class MockDialogService implements DialogService {
 	}
 
 	@Override
-	public void fileDownload(final Shell shell, final Path path) {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
 	public List<ToolboxFileFilter> getCsvFileFilters() {
 		throw new UnsupportedOperationException();
 	}
@@ -188,7 +183,7 @@ public class MockDialogService implements DialogService {
 
 	@Override
 	public Optional<Path> openFileDialog(final Shell shell,
-			final List<ToolboxFileFilter> filters) {
+			final List<ToolboxFileFilter> filters, Optional<Path> lastSelectedFile) {
 		return openFileDialogHandler.apply(filters);
 	}
 
@@ -222,12 +217,6 @@ public class MockDialogService implements DialogService {
 	@Override
 	public void reportSavedSession(final Shell shell,
 			final IModelSession session) {
-		throw new UnsupportedOperationException();
-	}
-
-	@Override
-	public Optional<Path> saveFileDialog(final Shell shell,
-			final List<ToolboxFileFilter> filters) {
 		throw new UnsupportedOperationException();
 	}
