+++ pom.xml
@@ -308,13 +308,6 @@
 							</environments>
 							<dependency-resolution>
 								<optionalDependencies>ignore</optionalDependencies>
-								<extraRequirements>
-									<requirement>
-										<type>eclipse-plugin</type>
-										<id>org.eclipse.set.rcp.compatibility</id>
-										<versionRange>0.0.0</versionRange>
-									</requirement>
-								</extraRequirements>
 							</dependency-resolution>
 						</configuration>
 					</plugin>
